requirements.txt
# === KAAN'IN RAG YAPAY ZEKA KÜTÜPHANELERİ ===
beautifulsoup4==4.12.3
langchain==0.2.7
langgraph==0.1.8
langchainhub==0.1.20
langchain-community==0.2.7
tavily-python==0.3.4
langchain-chroma==0.1.2
langchain-openai==0.1.16
pypdf

# === DİLA'NIN VERİTABANI KÜTÜPHANELERİ ===
SQLAlchemy==2.0.45
psycopg2-binary
passlib==1.7.4
bcrypt==3.2.0
python-jose==3.5.0
httpx

# === ORTAK ALTYAPI ===
fastapi
uvicorn[standard]
gunicorn
python-dotenv
python-multipart